Move streamcore package to streamutil
