+++ streamclient/inmemclient/consumer.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/blendle/go-streamprocessor/stream"
 	"github.com/blendle/go-streamprocessor/streamconfig"
-	"github.com/blendle/go-streamprocessor/streamcore"
+	"github.com/blendle/go-streamprocessor/streamutil"
 	"go.uber.org/zap"
 )
 
@@ -46,7 +46,7 @@ func NewConsumer(options ...func(*streamconfig.Consumer)) (stream.Consumer, erro
 	// configuration flag. If the auto-error functionality is disabled, the user
 	// needs to manually listen to the `Errors()` channel and act accordingly.
 	if consumer.c.HandleErrors {
-		go streamcore.HandleErrors(consumer.errors, consumer.logger.Fatal)
+		go streamutil.HandleErrors(consumer.errors, consumer.logger.Fatal)
 	}
 
 	// We start a goroutine to consume any messages currently stored in the inmem
@@ -64,7 +64,7 @@ func NewConsumer(options ...func(*streamconfig.Consumer)) (stream.Consumer, erro
 	// configuration flag.
 	if consumer.c.HandleInterrupt {
 		consumer.signals = make(chan os.Signal, 1)
-		go streamcore.HandleInterrupts(consumer.signals, consumer.Close, consumer.logger)
+		go streamutil.HandleInterrupts(consumer.signals, consumer.Close, consumer.logger)
 	}
 
 	return consumer, nil
@@ -79,7 +79,7 @@ func (c *Consumer) Messages() <-chan stream.Message {
 // Errors returns the read channel for the errors that are returned by the
 // stream.
 func (c *Consumer) Errors() <-chan error {
-	return streamcore.ErrorsChan(c.errors, c.c.HandleErrors)
+	return streamutil.ErrorsChan(c.errors, c.c.HandleErrors)
 }
 
 // Ack is a no-op implementation to satisfy the stream.Consumer interface.
