+++ streamclient/inmemclient/producer.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/blendle/go-streamprocessor/stream"
 	"github.com/blendle/go-streamprocessor/streamconfig"
-	"github.com/blendle/go-streamprocessor/streamcore"
+	"github.com/blendle/go-streamprocessor/streamutil"
 	"go.uber.org/zap"
 )
 
@@ -47,7 +47,7 @@ func NewProducer(options ...func(*streamconfig.Producer)) (stream.Producer, erro
 	// configuration flag. If the auto-error functionality is disabled, the user
 	// needs to manually listen to the `Errors()` channel and act accordingly.
 	if producer.c.HandleErrors {
-		go streamcore.HandleErrors(producer.errors, producer.logger.Fatal)
+		go streamutil.HandleErrors(producer.errors, producer.logger.Fatal)
 	}
 
 	// We listen to the produce channel in a goroutine. Every message delivered to
@@ -64,7 +64,7 @@ func NewProducer(options ...func(*streamconfig.Producer)) (stream.Producer, erro
 	// configuration flag.
 	if producer.c.HandleInterrupt {
 		producer.signals = make(chan os.Signal, 1)
-		go streamcore.HandleInterrupts(producer.signals, producer.Close, producer.logger)
+		go streamutil.HandleInterrupts(producer.signals, producer.Close, producer.logger)
 	}
 
 	return producer, nil
@@ -78,7 +78,7 @@ func (p *Producer) Messages() chan<- stream.Message {
 // Errors returns the read channel for the errors that are returned by the
 // stream.
 func (p *Producer) Errors() <-chan error {
-	return streamcore.ErrorsChan(p.errors, p.c.HandleErrors)
+	return streamutil.ErrorsChan(p.errors, p.c.HandleErrors)
 }
 
 // Close closes the producer connection. This function blocks until all messages
