+++ streamclient/kafkaclient/consumer.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/blendle/go-streamprocessor/stream"
 	"github.com/blendle/go-streamprocessor/streamconfig"
-	"github.com/blendle/go-streamprocessor/streamcore"
+	"github.com/blendle/go-streamprocessor/streamutil"
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"go.uber.org/zap"
 )
@@ -53,7 +53,7 @@ func NewConsumer(options ...func(*streamconfig.Consumer)) (stream.Consumer, erro
 	// configuration flag. If the auto-error functionality is disabled, the user
 	// needs to manually listen to the `Errors()` channel and act accordingly.
 	if consumer.c.HandleErrors {
-		go streamcore.HandleErrors(consumer.errors, consumer.logger.Fatal)
+		go streamutil.HandleErrors(consumer.errors, consumer.logger.Fatal)
 	}
 
 	// We start a goroutine to consume any messages being delivered to us from
@@ -71,7 +71,7 @@ func NewConsumer(options ...func(*streamconfig.Consumer)) (stream.Consumer, erro
 	// configuration flag.
 	if consumer.c.HandleInterrupt {
 		consumer.signals = make(chan os.Signal, 1)
-		go streamcore.HandleInterrupts(consumer.signals, consumer.Close, consumer.logger)
+		go streamutil.HandleInterrupts(consumer.signals, consumer.Close, consumer.logger)
 	}
 
 	return consumer, nil
@@ -86,7 +86,7 @@ func (c *Consumer) Messages() <-chan stream.Message {
 // Errors returns the read channel for the errors that are returned by the
 // stream.
 func (c *Consumer) Errors() <-chan error {
-	return streamcore.ErrorsChan(c.errors, c.c.HandleErrors)
+	return streamutil.ErrorsChan(c.errors, c.c.HandleErrors)
 }
 
 // Ack acknowledges that a message was processed. See `Consumer.storeOffset` for
